requirements.txt
discord.py
flask
requests
googlesearch-python
deep-translator
google-search
beautifulsoup4
python-dotenv